pyproject.toml
[project]
name = "ollama-cli"
version = "0.1.0"
description = "Full-featured AI coding assistant powered by Ollama with multi-provider support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Mohammed Uthmaan Al Hāshimī"}]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "python-semantic-release>=9.0.0",
]

dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
]

[project.scripts]
ollama = "cmd.root:main"
ollama-cli = "cmd.root:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cmd", "api", "model", "server", "runner"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.semantic_release]
branch = "main"
version_toml = "pyproject.toml:project.version"
version_source = "commit"
changelog_file = "CHANGELOG.md"
tag_format = "v{version}"
tag_commit = true
commit_parser = "angular"
commit_author_token = "${GH_TOKEN}"
commit_message = "chore(release): {version}\n\nAutomatically generated release notes."
build_command = "uv build --release"
upload_to_pypi = false
upload_to_release = true
remove_dist = false
assets = []